nueva estructura en BD y CRUD funcionando

+++ GuessPaint-Back/src/service/wordCategory.service.ts
@@ -1,27 +1,61 @@
 import { WordCategoryRepository } from "../repository/wordCategory.repository";
 import { WordCategory } from "../entity/WordCategory.entity";
+import { WordRepository } from "../repository/word.repository";
+import { CategoryRepository } from "../repository/category.repository";
+import { WordCategoryResponse } from "../dto/wordCategory.dto";
 
 export class WordCategoryService {
-    private wordCategoryRepository: WordCategoryRepository = new WordCategoryRepository();
+    private wordCategoryRepository: WordCategoryRepository ;
+    private wordRepository: WordRepository ;
+    private categoryRepository: CategoryRepository ;
 
-    public async findByCategoryId(category_id: string) {
-        return await this.wordCategoryRepository.findByCategoryId(category_id);
+    constructor(){
+        this.wordCategoryRepository = new WordCategoryRepository();
+        this.wordRepository = new WordRepository();
+        this.categoryRepository = new CategoryRepository();
     }
 
-    public async findByWordId(word_id: string) {
-        return await this.wordCategoryRepository.findByWordId(word_id);
+    public async save(wordCategory: WordCategoryResponse): Promise<WordCategory> {
+        const responseByIdWord = await this.wordRepository.findById(wordCategory.wordId);
+        const responseByIdCategory = await this.categoryRepository.findById(wordCategory.categoryId);
+        wordCategory.wordId = responseByIdWord.id;
+        wordCategory.categoryId = responseByIdCategory.id;
+        const wc_id= parseInt(responseByIdWord.id.toString() + responseByIdCategory.id.toString());
+        wordCategory.id = wc_id;
+
+        console.log(responseByIdWord.id)
+        console.log(responseByIdCategory.id)
+        console.log(wc_id)
+        console.log(wordCategory)
+
+
+
+        if(!responseByIdWord || !responseByIdCategory){
+            throw new Error('Word or Category not found');
+        }
+        return await this.wordCategoryRepository.save(wordCategory);
     }
 
-    public async getAll() {
+    public async findById(id: number): Promise<WordCategory | undefined> {
+        const wordCategory = await this.wordCategoryRepository.findById(id);
+        if (!wordCategory) {
+            throw new Error('WordCategory not found');
+        }
+        return wordCategory;
+    }
+
+    public async getAll(): Promise<WordCategory[]> {
         return await this.wordCategoryRepository.getAll();
     }
 
-    public async save(body: any) {
-        return await this.wordCategoryRepository.save(body);
-    }
-
-    public async delete(id: string) {
+    public async delete(id: number) {
+        const wordCategoryExist = await this.wordCategoryRepository.findById(id);
+        if (!wordCategoryExist) {
+            throw new Error('WordCategory not found');
+        }
+        else{
         return await this.wordCategoryRepository.delete(id);
+        }
     }
 
 }
